cleanup vault utils
