+++ e2e/examples/vault.rs
@@ -38,8 +38,8 @@ use valence_chain_client_utils::{
 use valence_e2e::utils::{
     ethereum::set_up_anvil_container,
     hyperlane::{
-        bech32_to_evm_bytes32, set_up_cw_hyperlane_contracts, set_up_eth_hyperlane_contracts,
-        set_up_hyperlane, HyperlaneContracts,
+        set_up_cw_hyperlane_contracts, set_up_eth_hyperlane_contracts, set_up_hyperlane,
+        HyperlaneContracts,
     },
     manager::{setup_manager, use_manager_init, ASTROPORT_LPER_NAME, ASTROPORT_WITHDRAWER_NAME},
     processor::tick_processor,
@@ -47,9 +47,9 @@ use valence_e2e::utils::{
         BaseAccount, MockERC20,
         ValenceVault::{self},
     },
-    ASTROPORT_PATH, DEFAULT_ANVIL_RPC_ENDPOINT, ETHEREUM_CHAIN_NAME, ETHEREUM_HYPERLANE_DOMAIN,
-    GAS_FLAGS, HYPERLANE_RELAYER_NEUTRON_ADDRESS, LOCAL_CODE_ID_CACHE_PATH_NEUTRON, LOGS_FILE_PATH,
-    NEUTRON_CONFIG_FILE, VALENCE_ARTIFACTS_PATH,
+    vault, ASTROPORT_PATH, DEFAULT_ANVIL_RPC_ENDPOINT, ETHEREUM_CHAIN_NAME,
+    ETHEREUM_HYPERLANE_DOMAIN, GAS_FLAGS, HYPERLANE_RELAYER_NEUTRON_ADDRESS,
+    LOCAL_CODE_ID_CACHE_PATH_NEUTRON, LOGS_FILE_PATH, NEUTRON_CONFIG_FILE, VALENCE_ARTIFACTS_PATH,
 };
 use valence_library_utils::liquidity_utils::AssetData;
 use valence_program_manager::{
@@ -58,7 +58,6 @@ use valence_program_manager::{
     program_config::ProgramConfig,
     program_config_builder::ProgramConfigBuilder,
 };
-use vault::vault_update;
 
 const EVM_ENCODER_NAMESPACE: &str = "evm_encoder_v1";
 const PROVIDE_LIQUIDITY_AUTHORIZATIONS_LABEL: &str = "provide_liquidity";
@@ -66,6 +65,7 @@ const WITHDRAW_LIQUIDITY_AUTHORIZATIONS_LABEL: &str = "withdraw_liquidity";
 const ASTROPORT_CONCENTRATED_PAIR_TYPE: &str = "concentrated";
 const SECONDS_IN_DAY: u64 = 86_400;
 
+/// macro for executing async code in a blocking context
 macro_rules! async_run {
     ($rt:expr, $($body:tt)*) => {
         $rt.block_on(async { $($body)* })
@@ -408,24 +408,13 @@ fn main() -> Result<(), Box<dyn Error>> {
     let deposit_amount = U256::from(500_000);
 
     info!("User depositing {deposit_amount}USDC tokens to vault...");
-    async_run!(rt, {
-        let client = eth_client.get_request_provider().await.unwrap();
-        let signed_tx = valence_vault
-            .deposit(deposit_amount, eth_user_acc)
-            .into_transaction_request()
-            .from(eth_user_acc);
-
-        match alloy::providers::Provider::send_transaction(&client, signed_tx).await {
-            Ok(resp) => {
-                let receipt = resp.get_receipt().await.unwrap();
-                info!("deposit success: {:?}", receipt.transaction_hash)
-            }
-            Err(e) => {
-                warn!("deposit failed: {:?}", e);
-                panic!("Failed to deposit initial tokens");
-            }
-        }
-    });
+    vault::deposit_to_vault(
+        &rt,
+        &eth_client,
+        *valence_vault.address(),
+        eth_user_acc,
+        deposit_amount,
+    )?;
 
     log_eth_balances(
         &eth_client,
@@ -440,17 +429,12 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     info!("performing vault update...");
 
-    let current_rate = async_run!(rt, {
-        eth_client
-            .query(valence_vault.redemptionRate())
-            .await
-            .unwrap()
-            ._0
-    });
+    let current_rate = vault::query_redemption_rate(*valence_vault.address(), &rt, &eth_client)._0;
+
     let netting_amount = U256::from(0);
 
     let withdraw_fee_bps = 1;
-    vault_update(
+    vault::vault_update(
         *valence_vault.address(),
         current_rate,
         withdraw_fee_bps,
@@ -464,45 +448,26 @@ fn main() -> Result<(), Box<dyn Error>> {
     vault::pause(*valence_vault.address(), &rt, &eth_client)?;
 
     info!("attempting to deposit to paused vault...");
-    info!("User depositing {deposit_amount}USDC tokens to vault...");
-    async_run!(rt, {
-        let client = eth_client.get_request_provider().await.unwrap();
-
-        let signed_tx = valence_vault
-            .deposit(deposit_amount, eth_user_acc)
-            .into_transaction_request()
-            .from(eth_user_acc);
-
-        match alloy::providers::Provider::send_transaction(&client, signed_tx).await {
-            Ok(resp) => warn!(
-                "deposit to paused vault succeeded: {:?}",
-                resp.get_receipt().await.unwrap().transaction_hash
-            ),
-            Err(_) => info!("failed to deposit; vault is paused"),
-        }
-    });
+    vault::deposit_to_vault(
+        &rt,
+        &eth_client,
+        *valence_vault.address(),
+        eth_user_acc,
+        deposit_amount,
+    )?;
 
     info!("resuming the vault...");
     vault::unpause(*valence_vault.address(), &rt, &eth_client)?;
 
     info!("attempting to deposit to active vault...");
-    info!("Depositing {deposit_amount} tokens to vault...");
-    async_run!(rt, {
-        let client = eth_client.get_request_provider().await.unwrap();
-
-        let signed_tx = valence_vault
-            .deposit(deposit_amount, eth_user_acc)
-            .into_transaction_request()
-            .from(eth_user_acc);
-
-        match alloy::providers::Provider::send_transaction(&client, signed_tx).await {
-            Ok(resp) => info!(
-                "deposit to paused vault succeeded: {:?}",
-                resp.get_receipt().await.unwrap().transaction_hash
-            ),
-            Err(_) => warn!("failed to deposit"),
-        }
-    });
+    vault::deposit_to_vault(
+        &rt,
+        &eth_client,
+        *valence_vault.address(),
+        eth_user_acc,
+        deposit_amount,
+    )
+    .unwrap();
 
     log_eth_balances(
         &eth_client,
@@ -512,8 +477,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         &deposit_acc_addr,
         &withdraw_acc_addr,
         &eth_user_acc,
-    )
-    .unwrap();
+    )?;
 
     async_run!(rt, {
         info!("minting some tokens into admin acc for withdraw prep...");
@@ -593,41 +557,26 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     info!("User initiates shares redeemal...");
 
-    let user_shares = async_run!(rt, {
-        eth_client
-            .query(valence_vault.balanceOf(eth_user_acc))
-            .await
-            .unwrap()
-            ._0
-    });
+    let user_shares =
+        vault::query_vault_balance_of(*valence_vault.address(), &rt, &eth_client, eth_user_acc)._0;
 
-    async_run!(rt, {
-        let client = eth_client.get_request_provider().await.unwrap();
-        let signed_tx = valence_vault
-            .redeem_0(user_shares, eth_user_acc, eth_user_acc, 10_000, true)
-            .into_transaction_request()
-            .from(eth_user_acc);
-        match alloy::providers::Provider::send_transaction(&client, signed_tx).await {
-            Ok(resp) => {
-                let receipt = resp.get_receipt().await.unwrap();
-                info!("redeem request response: {:?}", receipt.transaction_hash);
-            }
-            Err(e) => warn!("redeem request error: {:?}", e),
-        }
-    });
+    vault::redeem(
+        *valence_vault.address(),
+        &rt,
+        &eth_client,
+        eth_user_acc,
+        user_shares,
+        10_000,
+        true,
+    )?;
 
-    async_run!(rt, {
-        let active_withdraws = eth_client
-            .query(valence_vault.hasActiveWithdraw(eth_user_acc))
-            .await
-            .unwrap();
-        info!("user active withdraws: {:?}", active_withdraws._0);
-        let user_withdraw_request = eth_client
-            .query(valence_vault.userWithdrawRequest(eth_user_acc))
-            .await
-            .unwrap();
-        info!("user active withdraw request: {:?}", user_withdraw_request);
-    });
+    let has_active_withdraw =
+        vault::addr_has_active_withdraw(*valence_vault.address(), &rt, &eth_client, eth_user_acc);
+    info!("user active withdraws: {:?}", has_active_withdraw._0);
+
+    let user_withdraw_request =
+        vault::addr_withdraw_request(*valence_vault.address(), &rt, &eth_client, eth_user_acc);
+    info!("user active withdraw request: {:?}", user_withdraw_request);
 
     log_eth_balances(
         &eth_client,
@@ -643,35 +592,15 @@ fn main() -> Result<(), Box<dyn Error>> {
     sleep(Duration::from_secs(3));
 
     info!("user attempts to finalize the withdrawal...");
-    async_run!(rt, {
-        let client = eth_client.get_request_provider().await.unwrap();
+    vault::complete_withdraw_request(*valence_vault.address(), &rt, &eth_client, eth_user_acc)?;
 
-        let signed_tx = valence_vault
-            .completeWithdraw(eth_user_acc)
-            .into_transaction_request()
-            .from(eth_user_acc);
+    let has_active_withdraw =
+        vault::addr_has_active_withdraw(*valence_vault.address(), &rt, &eth_client, eth_user_acc);
+    info!("user active withdraws: {:?}", has_active_withdraw._0);
 
-        match alloy::providers::Provider::send_transaction(&client, signed_tx).await {
-            Ok(resp) => {
-                let receipt = resp.get_receipt().await.unwrap();
-                info!(
-                    "complete withdrawal request response: {:?}",
-                    receipt.transaction_hash
-                );
-            }
-            Err(e) => warn!("complete withdrawal request error: {:?}", e),
-        };
-        let active_withdraws = eth_client
-            .query(valence_vault.hasActiveWithdraw(eth_user_acc))
-            .await
-            .unwrap();
-        info!("user active withdraws: {:?}", active_withdraws._0);
-        let user_withdraw_request = eth_client
-            .query(valence_vault.userWithdrawRequest(eth_user_acc))
-            .await
-            .unwrap();
-        info!("user active withdraw request: {:?}", user_withdraw_request);
-    });
+    let user_withdraw_request =
+        vault::addr_withdraw_request(*valence_vault.address(), &rt, &eth_client, eth_user_acc);
+    info!("user active withdraw request: {:?}", user_withdraw_request);
 
     log_eth_balances(
         &eth_client,
@@ -730,398 +659,6 @@ fn log_eth_balances(
     });
 
     Ok(())
-}
-
-pub mod vault {
-    use std::{error::Error, str::FromStr};
-
-    use alloy::{
-        primitives::{Address, U256},
-        sol_types::SolValue,
-    };
-    use log::{info, warn};
-    use valence_chain_client_utils::{
-        ethereum::EthereumClient,
-        evm::{base_client::EvmBaseClient as _, request_provider_client::RequestProviderClient},
-    };
-    use valence_e2e::utils::{
-        solidity_contracts::{
-            BaseAccount, LiteProcessor,
-            MockERC20::{self},
-            ValenceVault::{self, FeeConfig, FeeDistributionConfig, VaultConfig},
-        },
-        NEUTRON_HYPERLANE_DOMAIN,
-    };
-
-    use crate::SECONDS_IN_DAY;
-
-    pub fn vault_update(
-        vault_addr: Address,
-        new_rate: U256,
-        withdraw_fee_bps: u32,
-        netting_amount: U256,
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-    ) -> Result<(), Box<dyn Error>> {
-        let eth_rp = async_run!(rt, eth_client.get_request_provider().await.unwrap());
-        let valence_vault = ValenceVault::new(vault_addr, &eth_rp);
-
-        let config = async_run!(rt, eth_client.query(valence_vault.config()).await.unwrap());
-        info!("pre-update vault config: {:?}", config);
-
-        let start_rate = async_run!(rt, {
-            eth_client
-                .query(valence_vault.redemptionRate())
-                .await
-                .unwrap()
-                ._0
-        });
-        let prev_max_rate = async_run!(rt, {
-            eth_client
-                .query(valence_vault.maxHistoricalRate())
-                .await
-                .unwrap()
-                ._0
-        });
-
-        let prev_total_assets = async_run!(rt, {
-            eth_client
-                .query(valence_vault.totalAssets())
-                .await
-                .unwrap()
-                ._0
-        });
-
-        info!("Vault start rate: {start_rate}");
-        info!("Vault current max historical rate: {prev_max_rate}");
-        info!("Vault current total assets: {prev_total_assets}");
-        info!(
-            "Updating vault with new rate: {new_rate}, withdraw fee: {withdraw_fee_bps}bps, netting: {netting_amount}"
-        );
-
-        let update_result = async_run!(rt, {
-            eth_client
-                .execute_tx(
-                    valence_vault
-                        .update(new_rate, withdraw_fee_bps, netting_amount)
-                        .into_transaction_request(),
-                )
-                .await
-        });
-
-        if let Err(e) = &update_result {
-            info!("Update failed: {:?}", e);
-            panic!("failed to update vault");
-        }
-
-        let new_redemption_rate = async_run!(rt, {
-            eth_client
-                .query(valence_vault.redemptionRate())
-                .await
-                .unwrap()
-                ._0
-        });
-        let new_max_historical_rate = async_run!(rt, {
-            eth_client
-                .query(valence_vault.maxHistoricalRate())
-                .await
-                .unwrap()
-                ._0
-        });
-
-        let new_total_assets = async_run!(rt, {
-            eth_client
-                .query(valence_vault.totalAssets())
-                .await
-                .unwrap()
-                ._0
-        });
-
-        let config = async_run!(rt, eth_client.query(valence_vault.config()).await.unwrap());
-        info!("Vault new config: {:?}", config);
-        info!("Vault new redemption rate: {new_redemption_rate}");
-        info!("Vault new max historical rate: {new_max_historical_rate}");
-        info!("Vault new total assets: {new_total_assets}");
-
-        assert_eq!(
-            new_redemption_rate, new_rate,
-            "Redemption rate should be updated to the new rate"
-        );
-
-        // Verify max historical rate was updated if needed
-        if new_rate > prev_max_rate {
-            assert_eq!(
-                new_max_historical_rate, new_rate,
-                "Max historical rate should be updated when new rate is higher"
-            );
-        } else {
-            assert_eq!(
-                new_max_historical_rate, prev_max_rate,
-                "Max historical rate should remain unchanged when new rate is not higher"
-            );
-        }
-
-        Ok(())
-    }
-
-    pub fn setup_vault_config(
-        accounts: &[Address],
-        eth_deposit_acc: Address,
-        eth_withdraw_acc: Address,
-    ) -> VaultConfig {
-        let fee_config = FeeConfig {
-            depositFeeBps: 0,        // No deposit fee
-            platformFeeBps: 1000,    // 10% yearly platform fee
-            performanceFeeBps: 2000, // 20% performance fee
-            solverCompletionFee: 0,  // No solver completion fee
-        };
-
-        let fee_distribution = FeeDistributionConfig {
-            strategistAccount: accounts[0], // Strategist fee recipient
-            platformAccount: accounts[1],   // Platform fee recipient
-            strategistRatioBps: 5000,       // 50% to strategist
-        };
-
-        VaultConfig {
-            depositAccount: eth_deposit_acc,
-            withdrawAccount: eth_withdraw_acc,
-            strategist: accounts[0],
-            fees: fee_config,
-            feeDistribution: fee_distribution,
-            depositCap: 0, // No cap (for real)
-            withdrawLockupPeriod: 1,
-            // withdrawLockupPeriod: SECONDS_IN_DAY, // 1 day lockup
-            maxWithdrawFeeBps: 100, // 1% max withdraw fee
-        }
-    }
-
-    pub fn pause(
-        vault_addr: Address,
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-    ) -> Result<(), Box<dyn Error>> {
-        async_run!(rt, {
-            let eth_rp = eth_client.get_request_provider().await.unwrap();
-            let valence_vault = ValenceVault::new(vault_addr, &eth_rp);
-
-            match eth_client
-                .execute_tx(valence_vault.pause().into_transaction_request())
-                .await
-            {
-                Ok(_) => info!("vault paused!"),
-                Err(_) => warn!("failed to pause the vault!"),
-            };
-
-            let packed_vals = eth_client
-                .query(valence_vault.packedValues())
-                .await
-                .unwrap();
-
-            assert!(packed_vals.paused, "vault should be paused");
-        });
-
-        Ok(())
-    }
-
-    pub fn unpause(
-        vault_addr: Address,
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-    ) -> Result<(), Box<dyn Error>> {
-        async_run!(rt, {
-            let eth_rp = eth_client.get_request_provider().await.unwrap();
-            let valence_vault = ValenceVault::new(vault_addr, &eth_rp);
-            match eth_client
-                .execute_tx(valence_vault.unpause().into_transaction_request())
-                .await
-            {
-                Ok(_) => info!("vault resumed!"),
-                Err(_) => warn!("failed to resume the vault!"),
-            };
-            let packed_vals = eth_client
-                .query(valence_vault.packedValues())
-                .await
-                .unwrap();
-
-            assert!(!packed_vals.paused, "vault should be unpaused");
-        });
-
-        Ok(())
-    }
-
-    pub fn query_vault_config(
-        vault_addr: Address,
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-    ) -> ValenceVault::configReturn {
-        let config = async_run!(rt, {
-            let eth_rp = eth_client.get_request_provider().await.unwrap();
-            let valence_vault = ValenceVault::new(vault_addr, &eth_rp);
-
-            eth_client.query(valence_vault.config()).await.unwrap()
-        });
-
-        info!("VAULT CONFIG config: {:?}", config);
-        config
-    }
-
-    pub fn query_vault_total_assets(
-        vault_addr: Address,
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-    ) -> ValenceVault::totalAssetsReturn {
-        let eth_rp = async_run!(rt, { eth_client.get_request_provider().await.unwrap() });
-        let valence_vault = ValenceVault::new(vault_addr, &eth_rp);
-
-        async_run!(rt, {
-            eth_client.query(valence_vault.totalAssets()).await.unwrap()
-        })
-    }
-
-    pub fn query_vault_total_supply(
-        vault_addr: Address,
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-    ) -> ValenceVault::totalSupplyReturn {
-        let eth_rp = async_run!(rt, { eth_client.get_request_provider().await.unwrap() });
-        let valence_vault = ValenceVault::new(vault_addr, &eth_rp);
-
-        async_run!(rt, {
-            eth_client.query(valence_vault.totalSupply()).await.unwrap()
-        })
-    }
-
-    pub fn query_vault_balance_of(
-        vault_addr: Address,
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-        addr: Address,
-    ) -> ValenceVault::balanceOfReturn {
-        let eth_rp = async_run!(rt, { eth_client.get_request_provider().await.unwrap() });
-        let valence_vault = ValenceVault::new(vault_addr, &eth_rp);
-
-        async_run!(rt, {
-            eth_client
-                .query(valence_vault.balanceOf(addr))
-                .await
-                .unwrap()
-        })
-    }
-
-    pub fn update() -> Result<(), Box<dyn Error>> {
-        // query both neutron and eth sides
-        // find netting amount
-        // update
-        Ok(())
-    }
-
-    pub fn setup_deposit_erc20(
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-    ) -> Result<Address, Box<dyn Error>> {
-        let eth_rp = async_run!(rt, { eth_client.get_request_provider().await.unwrap() });
-
-        info!("Deploying ERC20s on Ethereum...");
-        let evm_vault_deposit_token_tx =
-            MockERC20::deploy_builder(&eth_rp, "TestUSDC".to_string(), "TUSD".to_string())
-                .into_transaction_request();
-
-        let evm_vault_deposit_token_rx = async_run!(rt, {
-            valence_chain_client_utils::evm::base_client::EvmBaseClient::execute_tx(
-                eth_client,
-                evm_vault_deposit_token_tx,
-            )
-            .await
-            .unwrap()
-        });
-
-        let valence_vault_deposit_token_address =
-            evm_vault_deposit_token_rx.contract_address.unwrap();
-
-        Ok(valence_vault_deposit_token_address)
-    }
-
-    pub fn setup_valence_account(
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-        admin: Address,
-    ) -> Result<Address, Box<dyn Error>> {
-        let eth_rp = async_run!(rt, { eth_client.get_request_provider().await.unwrap() });
-
-        info!("Deploying base account on Ethereum...");
-        let base_account_tx =
-            BaseAccount::deploy_builder(&eth_rp, admin, vec![]).into_transaction_request();
-
-        let base_account_rx = async_run!(rt, {
-            eth_client
-                .execute_tx(base_account_tx.clone())
-                .await
-                .unwrap()
-        });
-
-        let base_account_addr = base_account_rx.contract_address.unwrap();
-
-        Ok(base_account_addr)
-    }
-
-    pub fn setup_lite_processor(
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-        admin: Address,
-        mailbox: &str,
-        authorization_contract_address: &str,
-    ) -> Result<Address, Box<dyn Error>> {
-        let eth_rp = async_run!(rt, { eth_client.get_request_provider().await.unwrap() });
-
-        let tx = LiteProcessor::deploy_builder(
-            &eth_rp,
-            valence_e2e::utils::hyperlane::bech32_to_evm_bytes32(authorization_contract_address)?,
-            Address::from_str(mailbox)?,
-            NEUTRON_HYPERLANE_DOMAIN,
-            vec![admin],
-        )
-        .into_transaction_request();
-
-        let lite_processor_rx = async_run!(rt, { eth_client.execute_tx(tx).await.unwrap() });
-
-        let lite_processor_address = lite_processor_rx.contract_address.unwrap();
-        info!("Lite Processor deployed at: {}", lite_processor_address);
-
-        Ok(lite_processor_address)
-    }
-
-    pub fn setup_valence_vault(
-        rt: &tokio::runtime::Runtime,
-        eth_client: &EthereumClient,
-        eth_accounts: &[Address],
-        admin: Address,
-        eth_deposit_acc: Address,
-        eth_withdraw_acc: Address,
-        vault_deposit_token_addr: Address,
-    ) -> Result<Address, Box<dyn Error>> {
-        let eth_rp = async_run!(rt, { eth_client.get_request_provider().await.unwrap() });
-
-        info!("deploying Valence Vault on Ethereum...");
-        let vault_config = setup_vault_config(eth_accounts, eth_deposit_acc, eth_withdraw_acc);
-
-        let vault_tx = ValenceVault::deploy_builder(
-            &eth_rp,
-            admin,                            // owner
-            vault_config.abi_encode().into(), // encoded config
-            vault_deposit_token_addr,         // underlying token
-            "Valence Test Vault".to_string(), // vault token name
-            "vTEST".to_string(),              // vault token symbol
-            U256::from(1e18), // placeholder, tbd what a reasonable value should be here
-        )
-        .into_transaction_request();
-
-        let vault_rx = async_run!(rt, { eth_client.execute_tx(vault_tx).await.unwrap() });
-
-        let vault_address = vault_rx.contract_address.unwrap();
-        info!("Vault deployed at: {vault_address}");
-
-        Ok(vault_address)
-    }
 }
 
 fn eth_side_setup(
